Made a change in UpgradeItem() function of both ComputerPlayer and HumanPlayer classes. Fully implemented PlayARound() function in Game class. Made a basic gameplay loop in main() function. Changes UpdateStats() function's parameter's type from 'long' to 'double' in all classes inherits GameObject class. Added ShowStats() function in DeckItem class. Added tonnes of System.out.println()s to show gameplay flow.

+++ src/GameObjects/GameObject.java
@@ -84,5 +84,5 @@ public abstract class GameObject
 
     public abstract double CalculateAttackValue(MasterScissor enemyMasterScissor);
 
-    public abstract int UpdateStats(long addToDurability, long addToLevelPoint);
+    public abstract int UpdateStats(double addToDurability, double addToLevelPoint);
 }
